feat: move `ConversationHeader` higher up the DOM tree
