+++ src/apps/messenger/Main.module.scss
@@ -1,5 +1,15 @@
 .Split {
+  position: relative;
   display: flex;
   width: 100%;
   gap: 8px;
+}
+
+.Header {
+  left: 0;
+  pointer-events: all !important;
+  position: absolute;
+  top: 16px;
+  width: calc(100% - 16px) !important;
+  z-index: 1000;
 }
