+++ src/apps/messenger/Main.tsx
@@ -10,6 +10,7 @@ import { DevPanelContainer } from '../../components/dev-panel/container';
 import { FeatureFlag } from '../../components/feature-flag';
 
 import styles from './Main.module.scss';
+import { ConversationHeader } from '../../components/messenger/chat/conversation-header/container';
 
 export interface Properties {
   context: {
@@ -34,6 +35,8 @@ export class Container extends React.Component<Properties> {
             <Sidekick />
 
             <div className={styles.Split}>
+              <ConversationHeader className={styles.Header} />
+
               <FeatureFlag featureFlag='enableChannels'>
                 <MessengerFeed />
               </FeatureFlag>
