+++ src/components/messenger/chat/index.test.tsx
@@ -8,9 +8,7 @@ import { MessageInput } from '../../message-input/container';
 import { Media } from '../../message-input/utils';
 import { LeaveGroupDialogContainer } from '../../group-management/leave-group-dialog/container';
 import { JoiningConversationDialog } from '../../joining-conversation-dialog';
-
-import { StoreBuilder, stubAuthenticatedUser, stubConversation, stubUser } from '../../../store/test/store';
-import { ConversationHeader } from './conversation-header';
+import { stubUser } from '../../../store/test/store';
 
 const mockSearchMentionableUsersForChannel = jest.fn();
 jest.mock('../../../platform-apps/channels/util/api', () => {
@@ -27,15 +25,9 @@ describe(DirectMessageChat, () => {
       activeConversationId: '1',
       leaveGroupDialogStatus: LeaveGroupDialogStatus.CLOSED,
       directMessage: { id: '1', otherMembers: [] } as any,
-      canLeaveRoom: false,
-      canEdit: false,
-      canAddMembers: false,
-      canViewDetails: false,
       isSecondarySidekickOpen: false,
-      isRoomMuted: false,
       sendMessage: () => null,
       onRemoveReply: () => null,
-      isCurrentUserRoomAdmin: false,
       isJoiningConversation: false,
       startAddGroupMember: () => null,
       startEditConversation: () => null,
@@ -83,15 +75,6 @@ describe(DirectMessageChat, () => {
     const wrapper = subject({ isJoiningConversation: false });
 
     expect(wrapper).not.toHaveElement(JoiningConversationDialog);
-  });
-
-  it('calls start add group member', async function () {
-    const startAddGroupMember = jest.fn();
-    const wrapper = subject({ startAddGroupMember });
-
-    wrapper.find(ConversationHeader).simulate('addMember');
-
-    expect(startAddGroupMember).toHaveBeenCalledOnce();
   });
 
   describe('leave group dialog', () => {
@@ -187,219 +170,6 @@ describe(DirectMessageChat, () => {
       await input.prop('getUsersForMentions')('bob');
 
       expect(mockSearchMentionableUsersForChannel).toHaveBeenCalledWith('5', 'bob', []);
-    });
-  });
-
-  describe('mapState', () => {
-    describe('canLeaveRoom', () => {
-      it('is false when only when one other member', () => {
-        const state = new StoreBuilder().withActiveConversation(
-          stubConversation({ otherMembers: [stubUser()], isSocialChannel: false })
-        );
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canLeaveRoom: false }));
-      });
-
-      it('is false when user is admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({
-              otherMembers: [stubUser(), stubUser()],
-              adminMatrixIds: ['current-user-matrix-id'],
-              isSocialChannel: false,
-            })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canLeaveRoom: false }));
-      });
-
-      it('is false when user is admin and conversation is social channel', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({
-              otherMembers: [stubUser(), stubUser()],
-              adminMatrixIds: ['current-user-matrix-id'],
-              isSocialChannel: true,
-            })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canLeaveRoom: false }));
-      });
-
-      it('is true when user is not admin and more than one other member', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({ otherMembers: [stubUser(), stubUser()], adminMatrixIds: ['other-user-matrix-id'] })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canLeaveRoom: true }));
-      });
-
-      it('is true when the conversation is a social channel and has multiple members', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(stubConversation({ otherMembers: [stubUser(), stubUser()], isSocialChannel: true }))
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canLeaveRoom: true }));
-      });
-
-      it('is false when the conversation is a social channel and has only one member', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(stubConversation({ otherMembers: [stubUser()], isSocialChannel: true }))
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canLeaveRoom: false }));
-      });
-    });
-
-    describe('canEdit', () => {
-      it('is false when one on one conversation', () => {
-        const state = new StoreBuilder().withActiveConversation(
-          stubConversation({ isOneOnOne: true, isSocialChannel: false })
-        );
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canEdit: false }));
-      });
-
-      it('is false when current user is not room admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({ isOneOnOne: false, adminMatrixIds: ['other-user-matrix-id'], isSocialChannel: false })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canEdit: false }));
-      });
-
-      it('is true when current user is room admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({ isOneOnOne: false, adminMatrixIds: ['current-user-matrix-id'], isSocialChannel: false })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canEdit: true }));
-      });
-
-      it('is true when current user is room moderator', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({ isOneOnOne: false, moderatorIds: ['current-user-id'], isSocialChannel: false })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ id: 'current-user-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canEdit: true }));
-      });
-
-      it('is true when the conversation is a social channel and user is an admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({
-              isOneOnOne: true,
-              adminMatrixIds: ['current-user-matrix-id'],
-              isSocialChannel: true,
-            })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canEdit: true }));
-      });
-
-      it('is false when the conversation is a social channel and current user is not room admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({ isOneOnOne: false, adminMatrixIds: ['other-user-matrix-id'], isSocialChannel: true })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canEdit: false }));
-      });
-    });
-
-    describe('canAddMembers', () => {
-      it('is false when one on one conversation', () => {
-        const state = new StoreBuilder().withActiveConversation(
-          stubConversation({ isOneOnOne: true, isSocialChannel: false })
-        );
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canAddMembers: false }));
-      });
-
-      it('is false when current user is not room admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({ isOneOnOne: false, adminMatrixIds: ['other-user-matrix-id'], isSocialChannel: false })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canAddMembers: false }));
-      });
-
-      it('is true when current user is room admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({ isOneOnOne: false, adminMatrixIds: ['current-user-matrix-id'], isSocialChannel: false })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canAddMembers: true }));
-      });
-
-      it('is true when the conversation is a social channel and user is an admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({
-              isOneOnOne: true,
-              adminMatrixIds: ['current-user-matrix-id'],
-              isSocialChannel: true,
-            })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canAddMembers: true }));
-      });
-
-      it('is false when the conversation is a social channel and current user is not room admin', () => {
-        const state = new StoreBuilder()
-          .withActiveConversation(
-            stubConversation({ isOneOnOne: false, adminMatrixIds: ['other-user-matrix-id'], isSocialChannel: true })
-          )
-          .withCurrentUser(stubAuthenticatedUser({ matrixId: 'current-user-matrix-id' }));
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canAddMembers: false }));
-      });
-    });
-
-    describe('canViewDetails', () => {
-      it('is false when one on one conversation', () => {
-        const state = new StoreBuilder().withActiveConversation(
-          stubConversation({ isOneOnOne: true, isSocialChannel: false })
-        );
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canViewDetails: false }));
-      });
-
-      it('is true when not a one on one conversation', () => {
-        const state = new StoreBuilder().withActiveConversation(
-          stubConversation({ isOneOnOne: false, isSocialChannel: false })
-        );
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canViewDetails: true }));
-      });
-
-      it('is true when the conversation is a social channel', () => {
-        const state = new StoreBuilder().withActiveConversation(
-          stubConversation({
-            isOneOnOne: false,
-            isSocialChannel: true,
-          })
-        );
-
-        expect(DirectMessageChat.mapState(state.build())).toEqual(expect.objectContaining({ canViewDetails: true }));
-      });
     });
   });
 });
