+++ src/components/messenger/chat/styles.scss
@@ -49,21 +49,6 @@ $title-bar-border-radius: 8px 8px 0px 0px;
     }
   }
 
-  &__header-gradient {
-    position: absolute;
-    width: 100%;
-    height: 64px;
-    z-index: 2;
-  }
-
-  &__header-position {
-    position: absolute;
-    width: 100%;
-    padding: 16px 16px 16px 0px;
-    box-sizing: border-box;
-    z-index: 3;
-  }
-
   &__footer-position {
     position: absolute;
     bottom: 0px;
